Massive overhaul to a ton of things. Have login and register modals. Have working login/register/logout functionality. Split out various bits of functionality! Everything is now in position for me to actually work on assignment 5. Also, ugh PHP.

+++ assign3/StatisticsViewModel.js
@@ -82,7 +82,7 @@ StatisticsViewModel = function() {
 		if(biggestKey == 0)
 			return "N/A";
 		
-		return parseInt(biggestKey).toFixed(2);
+		return parseFloat(biggestKey).toFixed(2);
 	});
